Tweaks code for empty lists
